fix: PHPStan level 7 — zero errors
- Type-narrow UpcomingHolidayWidget Holiday|Builder → Holiday|null
- Fix SmartShoppingList pivot property types and max() on list
- Fix SocialPostForm Product|Collection → @var Product|null
- Rewrite StripeConnectWebhookController to use data_get() for
  object/array property access
- Type-narrow Tenant::find() with @var Tenant|null
- Cast float properties from sum() results (FinanceSummary)
- Cast int return from sum() (ReferralProgram)
- Fix QrCode HtmlString|string print/cast issues
- Widen CreateOrder @param to array<string, mixed>
- Fix ucfirst on enum values, remove unnecessary null coalescing
- Remove unused argument.unresolvableType ignore rule (12 → 11)

PHPStan level 7: 96 → 0 errors

+++ app/Filament/Pages/ReferralProgram.php
@@ -57,7 +57,7 @@ class ReferralProgram extends Page
 
     public function getMonthsEarned(): int
     {
-        return Referral::query()->where('referrer_tenant_id', tenant()->id)
+        return (int) Referral::query()->where('referrer_tenant_id', tenant()->id)
             ->where('status', ReferralStatus::Rewarded)
             ->sum('reward_months');
     }
